Moved micro-features to respective packages
